feat(ui): lock left showcase panel size and implement cascading overlap accordion carousel with live hover preview

+++ src/DubbedUp.Godot/AudioPlayback/MenuMusicController.cs
@@ -103,10 +103,10 @@ public partial class MenuMusicController : Node
         if (_player is null || !_player.Playing) return;
 
         // Loop playback when track finishes
-        var streamLen = (float)(_player.Stream?.GetLength() ?? 0.0);
-        if (streamLen > 1.0f && _player.GetPlaybackPosition() >= streamLen - 0.1f)
+        var streamLen = _player.Stream?.GetLength() ?? 0.0;
+        if (streamLen > 1.0 && _player.GetPlaybackPosition() >= streamLen - 0.1)
         {
-            _player.Seek(0.0f);
+            _player.Seek(0.0);
         }
 
         // Beat pulse calculation for UI animations
@@ -163,3 +163,4 @@ public partial class MenuMusicController : Node
         }
     }
 }
+
